spikes/mcp-client-starter/mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.64.0",
    "fastmcp>=2.11.3",
    "mcp>=1.13.0",
    "python-dotenv>=1.1.1",
]
